+++ apps/backend/src/app/api/latest/emails/render-email/route.tsx
@@ -1,8 +1,7 @@
-import { EMAIL_THEMES, renderEmailWithTheme } from "@/lib/email-themes";
+import { renderEmailWithTemplate, renderEmailWithTheme } from "@/lib/email-themes";
 import { createSmartRouteHandler } from "@/route-handlers/smart-route-handler";
 import { KnownErrors } from "@stackframe/stack-shared/dist/known-errors";
-import { yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
-import { getEnvVariable } from "@stackframe/stack-shared/dist/utils/env";
+import { adaptSchema, yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
 import { captureError, StackAssertionError, StatusError } from "@stackframe/stack-shared/dist/utils/errors";
 
 
@@ -15,10 +14,12 @@ export const POST = createSmartRouteHandler({
   request: yupObject({
     auth: yupObject({
       type: yupString().oneOf(["admin"]).defined(),
-    }).nullable(),
+      tenancy: adaptSchema.defined(),
+    }).defined(),
     body: yupObject({
-      theme: yupString().oneOf(Object.keys(EMAIL_THEMES) as (keyof typeof EMAIL_THEMES)[]).defined(),
-      preview_html: yupString().defined(),
+      theme_id: yupString().defined(),
+      preview_html: yupString(),
+      template_id: yupString(),
     }),
   }),
   response: yupObject({
@@ -28,11 +29,27 @@ export const POST = createSmartRouteHandler({
       html: yupString().defined(),
     }).defined(),
   }),
-  async handler({ body }) {
-    if (!getEnvVariable("STACK_FREESTYLE_API_KEY")) {
-      throw new StatusError(500, "STACK_FREESTYLE_API_KEY is not set");
+  async handler({ body, auth: { tenancy } }) {
+    const themeList = tenancy.completeConfig.emails.themeList;
+    const templateList = tenancy.completeConfig.emails.templateList;
+    if (!Object.keys(themeList).includes(body.theme_id)) {
+      throw new StatusError(400, "No theme found with given id");
     }
-    const result = await renderEmailWithTheme(body.preview_html, body.theme);
+    if ((body.preview_html && body.template_id) || (!body.preview_html && !body.template_id)) {
+      throw new StatusError(400, "Exactly one of preview_html or template_id must be provided");
+    }
+    if (body.template_id && !Object.keys(templateList).includes(body.template_id)) {
+      throw new StatusError(400, "No template found with given id");
+    }
+    const theme = themeList[body.theme_id];
+    let result;
+    if (body.template_id) {
+      const template = templateList[body.template_id];
+      result = await renderEmailWithTemplate(template.tsxSource, theme.tsxSource, { projectDisplayName: tenancy.project.display_name });
+    }
+    else {
+      result = await renderEmailWithTheme(body.preview_html!, theme.tsxSource);
+    }
     if ("error" in result) {
       captureError('render-email', new StackAssertionError("Error rendering email with theme", { result }));
       throw new KnownErrors.EmailRenderingError(result.error);