+++ docs/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @stackframe/stack-docs
 
+## 2.8.23
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.23
+  - @stackframe/stack@2.8.23
+
+## 2.8.22
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.22
+  - @stackframe/stack@2.8.22
+
+## 2.8.21
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack@2.8.21
+
+## 2.8.20
+
+### Patch Changes
+
+- @stackframe/stack@2.8.20
+
 ## 2.8.19
 
 ### Patch Changes